// src/types/BitgetFF.ts
export interface BitgetFF {
  open_at: string;
  close_at: string;
  long_close_at: string;
  short_close_at: string;
  cex: string;
  act_name: string;
  symbol: string;
  deal_usdt: number;
  lever: number;
  long_size: number;
  short_size: number;
  status: number;
  up_status: number;
  px: number;
  long_trigger_px: number;
  long_fx_trigger_px: number;
  short_trigger_px: number;
  short_fx_trigger_px: number;
  pnl: number;
  pnl_ratio: number;
  long_order_id: string;
  long_client_order_id: string;
  long_tracking_no: string;
  long_open_usdt: number;
  long_open_px: number;
  long_achieved_pl: number;
  long_close_px: number;
  long_fee: number;
  short_order_id: string;
  short_client_order_id: string;
  short_tracking_no: string;
  short_open_usdt: number;
  short_open_px: number;
  short_achieved_pl: number;
  short_close_px: number;
  short_fee: number;
}